requirements.txt
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-spreadsheet==0.0.6
googleapis-common-protos==1.53.0
pandas==1.1.4
requests==2.25.0
lxml==4.5.1
APScheduler==3.0.0
beautifulsoup4==4.9.3
html5lib==1.1